Restructure content block file structure
