+++ frontend/src/components/module-viewer/content/blocks/index.ts
@@ -1,8 +1,7 @@
 import ButtonBlock from "./ButtonBlock";
-import TextBlock from "./TextBlock";
 import ImageBlock from "./ImageBlock";
+import TextBlock from "./TextBlock";
 import VideoBlock from "./VideoBlock";
 import HeadingBlock from "./HeadingBlock";
 
 export { ButtonBlock, TextBlock, ImageBlock, VideoBlock, HeadingBlock };
-// TODO: Just export ContentBlock
